backend/requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image Processing
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# AI Models - HuggingFace Transformers
transformers==4.37.0
torch==2.2.0
torchvision==0.17.0
accelerate==0.26.0
sentencepiece==0.1.99

# Groq API for LLM calls
groq==0.5.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0

# Music RAG
h5py==3.10.0
sentence-transformers==2.3.1
chromadb==0.4.22